[cli] move `add-to-gitignore.ts` to `src/util` On branch fix/env-pull-add-extra-rule-to-ignore Changes to be committed: 	renamed:    src/util/link/add-to-gitignore.ts -> src/util/add-to-gitignore.ts 	modified:   packages/cli/src/util/projects/link.ts 	modified:   src/commands/env/pull.ts 	modified:   src/util/link/repo.ts
