+++ packages/cli/src/util/add-to-gitignore.ts
@@ -1,7 +1,7 @@
 import os from 'os';
 import { join } from 'path';
 import { readFile, writeFile } from 'fs-extra';
-import { VERCEL_DIR } from '../projects/link';
+import { VERCEL_DIR } from './projects/link';
 
 export async function addToGitIgnore(path: string, ignore = VERCEL_DIR) {
   let isGitIgnoreUpdated = false;
